requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
